requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
langgraph
pydantic
requests
streamlit
python-dotenv
firebase-admin